Placed `unchanged` function in `expr.py`
