+++ sympy/polys/tests/test_rootoftools.py
@@ -16,7 +16,7 @@ from sympy import (
 )
 
 from sympy.utilities.pytest import raises, slow
-from sympy.utilities.misc import unchanged
+from sympy.core.expr import unchanged
 from sympy.core.compatibility import range
 
 from sympy.abc import a, b, x, y, z, r
