+++ sympy/utilities/misc.py
@@ -8,7 +8,6 @@ import re as _re
 import struct
 from textwrap import fill, dedent
 from sympy.core.compatibility import get_function_name, range, as_int
-from sympy.core.sympify import sympify
 
 
 
@@ -441,24 +440,3 @@ def ordinal(num):
     else:
         suffix = 'th'
     return str(n) + suffix
-
-
-def unchanged(func, *args):
-    """Return True if `func` applied to the `args` is unchanged.
-    Can be used instead of `assert foo == foo`.
-
-    Examples
-    ========
-
-    >>> from sympy.utilities.misc import unchanged
-    >>> from sympy.functions.elementary.trigonometric import cos
-    >>> from sympy.core.numbers import pi
-
-    >>> unchanged(cos, 1)  # instead of assert cos(1) == cos(1)
-    True
-
-    >>> unchanged(cos, pi)
-    False
-    """
-    f = func(*args)
-    return f.func == func and f.args == tuple([sympify(a) for a in args])
